05-blueprints/multitenant-agentic-platform/src/cdk_requirements.txt
aws-cdk-lib>=2.170.0
constructs>=10.0.0
